Editorial amendments to README
The aim of this commit is to make the language in the README a little
bit more accessible to absolute newcomers who may be wondering what
TRIL is.  To us after 20+ years of familiarity with J9, saying
"language" vs "representation" does not pose a source of confusion,
but a newcomer may be baffled whether "Testarossa intermediate
representation" and "TRIL" are the same thing.