Generate prologue and epilogues *after* instruction selection and RA
This commit changes compilation sequence to generate prologue and
epilogue(s) after instruction selection and register allocation phases.

This is necessary because only after that we know which registers we
need to save (reload) in prologue (epilogue).

In order to be able to do this, during instruction selection when
evaluating `?return` opcode, we insert a pseudo instruction `TRLeave`
which marks the exist point. Later in `#generateEpilogues` we scan
through instructions and inject real epilogue.