Fix automatic creation of `GNUmakefile.local`
The problem was that we need to include `GNUmakefile.local` early on
(before other makefiles) as it contains customizations. The default
template for it is howver defined in those makefiles, so attempt to build
it would cause a loop. GNU make detects that and abort it's creation.

To circumvent that, we include `GNUmakefile.local` only if it exists
and recreate it as part of `all` target. Not ideal, but better than
forcing user to