Introduce new CFG object to manage function control flow graph
This commit introduces a new class - `TRCFG` that encapslutates function's
control flow graph (represented as a directed graph of `TRILBlock`s connected
through successor links) and provide functions to validate the CFG and
traverse basic blocks.

This is (will be) needed to allow TRIL transformations that modify
the CFG and to push linearization of CFG down the road (just before
codegen).