Remove reference to parent from TRIL node
This commit removes reference to node's parent in order to support node
sharing (that is, two parent may share same child node so single parent
reference makes no sense).

Consequently, `TRILNode >> replaceWith:` cannot be easily implemented
anymore so it was removed and `TRILSimplifier` had to be updated to
use `TRILNode >> replace:with:`.

While at it, also fix the case when `TRILSimplifier` simplifies a treetop
(in which case block had to be updated accordingly). This case was not
handled at all before this commit.