Ensure that every virtual register is assigned only once
This commit makes sure that - at TRIL level - virtual registers are
assigned only once. Also, it makes sure that tree-tops do not produce
a value (that is, do not assign a virtual registers) as it cannot be
read.

This in turn required changes in `?load` / `?store` implementation - we
can no longer reuse previously used virtual registers as that would mean
(in theory) assigning them more than once. This was essentially a hack to
improve quality of generated code.

This commit reverts that logic and thus less efficient code is generated.
TRIL-level optimizations like `?load` commoning may improve the code, but
that's left for subsequent commit.