Introduce an utility class `TRSharedPool` - a base superclass for Tinyrossa pools
Pharo does not guarantee that that a shared pool is initialized before
any of the class using it is initialized. This is recurring problem but
nobody seems to bother fixing it.

This is especially bothersome in Tinyrossa because of its extensive use
of pools. To solve this problem, this commit introduces a base superclass
for all of its pools which provide `#initializeIfNotAlready`. This method
is intended to be called from pool user's own `#initialize` to make sure
pool is initialized (should it need any of pool values).

This is more of a hack, really.