Reimplement `TRDataType >> #loadOpcode` and `#storeOpcode`
This commit changes implementation of `#loadOpcode` and `#storeOpcode`
to use "opcode tables" rather than overriding the method on in each
datatype. The tables are stored in (new) pool `TROpcodeTables`.

While this might seem to be 'against OOP' and dispatch should be used
(as it was), this implementation allows to use nested tables to implement
more complex opcode lookups (like "compare opcode for greater or equal for
two Int64s").