Remove register from `TRRegisterMappedSymbol`
Commit ae455cd02 removed the logic that reused the (virtual) register
keeping the value of an parameter or automatic. However, it did not
remove code in `TRCodeEvaluator` which was responsible for clearing
that register from symbol to prevent (virtual) register reuse across
extended basic blocks.

Once this has been removed, there's no longer need for `TRRegisterMappedSymbol`
to have an register, so this commit removes it. Funny that something
called "register-mapped symbol" has no register.