Make sure each node is visited only once by TRIL rewriting optimizations
This commit makes sure each node is visited only once when rewriting TRIL.
This fixes a subtle bug in `TRILCommoner`. Imagine following TRIL:

    n3  istore R
    n4    icall "F"
    n5      iload R

    n6  istore S
    n4    ==> icall "F"

When processing child `n5` of `n6` (third store) we *must not* traverse
children on `n4` because otherwise we will get invalid TRIL!

This is because load-after-store elimination the commoner is trying to
do in order to eliminate stores into automatics (and automatics itself),

For the above TRIL (if it would process node multiple times), it would
traverse children of `n4` as part of processing `n6` and replace `n5`
with 'last' store to R at this point which is R, creating:

    n3  istore R
    n4    icall "F"
    n4      ==> icall "F"

    n6  istore S
    n4    ==> icall "F"
    n4      ==> icall "F"

This create a cycle in TRIL making it invalid. Making sure each node is
processed only once avoid this problem.