Simplify conditional branches where condition result is known statically
This commit add a simplification step that simplifies conditional
branches when the condition result is known in advance. For example:

    if?cmp?? <label>
	  ?const <v1>
	  ?const <v2>

is simplified to either

    goto <label>

(if condition always holds) or removed altogether if it does not.

Note, that this simplification changes the CFG so it may open up
opportunities for further optimizations. Because of this, this commit
changed the default optimization pipeline to run all optimizations
twice.