Prune CFG after simplification pass
This commit prunes the CFG after CFG simplification pass by discarding
all unreachable blocks. Discarding a block consist of removing all its
treetops and removing the block from CFG's internal list of blocks.

The reason for this is that this may drop use-count on symbols used by
discarded blocks and therefore open up an opportunity to further eliminate
(unnecessary) stores and/or not even reserving space for unused symbol.

Alternative (and maybe) better solution might be to make `TRILBlock`
use-counted and when blocks use-count drops to zero, decrement use count
on all its treetops. This is left as future work.