Allow "branching" node to be in the "middle" of a block
This commit relaxes rules on block contents regarding branches. Prior this
commit branch node must have been the last one in block's treetops.

This posed problem with "check" nodes, in particular with `overflowchk`.
The standard sequence to perform overflow check is:

         BB1:
    N01  overflowchk handler
    N02    <operation>
    N03      <child1>
    N04      <child2>
    N03    ==> <child1>
    N04    ==> <child2>
    N05  store temp
    N02    ==> <operation>

         BB2:
    N06    goto BB3

         handler:
    N??     ...

         BB3:
    N??     ...

If we mandate the `oveflowchk` node (N01) being the last, there would
be no way to hold on the result of `<operation>` since nodes cannot be
shared across block boundary.

To solve this, this commit relaxes the rules a bit and allows branches
and checks (so called "branching nodes") to be in the "middle" of a block,
that is, they no longer need to be the last in the block. This is seems
to be in line with original Testarossa design.

Since Tinyrossa does support at most two successors per block, each block
may contain at most one "branching node". This is enforced and checked
for at multiple places to avoid creation of such blocks in a first place.