Fix copying node locations when replacing node by another
Prior this commit, new node was assigned location of the old node. This
works fine for some transformations like constant folding or similar.

It does not work quite as well when replacing nodes in order to share
them (so-called "commoning").

This commit tries to "fix" that by setting new node's location only if
not already set.

Perhaps, this is won't work in all cases and we'd have to resort in doing
this explicitly in each of the transformations. Time will show.