Propagate location to newly created `#orphan` builders
This commit propagates current location to orphan builders (if any set).

Just like in previous commit, this may not work well in all cases and we
may resort to make it manually after creating an orphan. Time will tell,
but for now, we try to save the front-end implementor form having to think
of it.