When fixing up registers, build fixup map manually
This commit creates register fixup map manually rather than depending on
fact that virtual register converts to concrete bitvector when real register
is allocated.

This is not only easier to understand but also makes it easier debug
RA because inserted spills / reloads / moves refer to virtual registers
so it is clear which (virtual) register is spilled.