RLSRA: keep track of all uses of a virtual register
This commit changes live interval structure to keep track of all
uses of a virtual registers rather than just its definition (start)
and last use (end). It also keeps track whether at given position
the register was defined (written to) or used (value read) or both.

This is a preparation for spilling / reloading when running out
of available registers as well as with deciding which register to spill
(based on how many times and how "far" it is given virtual register
defd/used).

Internally, both def and use positions are are stored single ordered
array: def position `p` is stored as `(p*2) + 1` and use position
as `(p*2)`.