Remove (old) linear scan and (even older) naive constraint solving register allocators
This commit removes `TRLinearScanRegisterAllocator` and `TRNaiveConstraintSolvingRegisterAllocator`
register allocators.

The main reason is that commit "Let the register allocator to handle
(virtual) register dependencies" moves the responsibility to move values
to argument registers and from return register(s) to register allocator
based on register dependencies.

Neither `TRLinearScanRegisterAllocator` and `TRNaiveConstraintSolvingRegisterAllocator`
can be easily updated to work with this change, so this commit removes
them.