RLSRA: handle pre and post dependencies via interval splitting
So far, handling pre and post register dependencies was very limited:
it could only handle dependencies on non-allocatable registers. This
made things simple - all that was needed was to move value to / from
required real register. Since dependency could be only on non-allocatable
registers, there was no way to introduce conflict.

The downside of this was that this limited a number of registers
available - for example argument / return registers could not be
allocated: what a waste of good registers!

This commit improves handling of dependencies leveraging support for
interval splitting.