RLSRA: correctly update register dependencies when splitting intervals
This commit fixes horrible bug in RSLRA causing arguments not being
moved to argument registers when the original interval was split.

This was because when updating instructions to use newly create v-reg
for just-split interval, register dependencies were not updated,
referring to original v-reg.

To fix this problem, this commit introduces mew method -
`#replaceVirtualRegistersUsing:` - to perform v-reg replacements which
updates not only register used by the instruction, but also those
referred-to in dependencies (and possibly other Tinyrossa-specific
metadata attached to an instruction).