Decouple label symbol from basic block
So far each label symbol (`TRLabelSymbol`) has been associated with
block (`TRILBlock`). This commits relaxes this association - every
label now may or may not be associated with block. This allows labels to
be also used in "local control flows" (local meaning within one
evaluator). It will also allow to unify labels with other symbol types
(such as function symbols or statics).

Technically, a `TRLabelSymbol` the association is still from symbol to
block. I have tried to make it the other way round, from block to it's
symbol, but that made code managing CFG bit ugly so I decided to keep it
like this for time being.