Anchor calls to a treetop when building TRIL using `TRILBuilder`
This commit changes the behavior of `TRILBuilder` w.r.t building calls.
Since calls (`icall`, `acall` and so on) are not treetops, they were
not added to the block. Therefore if the result of the call (return
value) was not used, the call was elided from built IL. This is wrong,
call may have side-effects so they must not be elided.

This commit changes the behavior of `TRILBuilder` so that each call is
automatically anchored using `treetop`. For example, following:

   builder istore: {
     builder icall: { 'get_some_int' } .
     'temp' }

used to produce:

   N002 istore 'temp'
   N001   icall 'get_some_int'

With this commit, it produces:

   N002 treetop
   N001   icall 'get_some_int'
   N003 istore 'temp'
   N001   ==> icall 'get_some_int'

If ever we really want to elide unnecessary calls (whateber that means)
it should be implemented as an optimization pass.