Add `TRTarget >> endian`
This commit adds `TRTarget >> endian` which can be used by various parts
when reading from / writing to (target) memory.

As of now, both RISC-V and POWER targets are hard-coded as little endian
even thought this is not always true. This may be fixes in the future.

The rest of Tinyrossa code should consult `TRTarget >> endian` wherever
necessary.