Add `TRTarget >> sizeInBytesOf:` replacing `TRDataType >> #sizeInBytes`
Commit b32bc8e added `TRDataType >> #sizeInBytes`. While it is
convenient and looks like a good OO design, it does not really work.
The size of `Address` (pointer) is architecture-dependent.

This commit deprecates `TRDataType >> #sizeInBytes` in favour of
`TRTarget >> sizeInBytesOf:` and updates users.

Note, that `TRDataType >> #sizeInBytes` has been kept there as
"should-not-implement" with comment to prevent future myself to make
this tempting mistake again.