RISC-V: save only used parameters in frame
This commit changes the RISC-V PSABI calling convention to not store
unused parameters into the frame in prologue. This not only keeps frame
smaller but more importantly avoids (expensive) memory write.

Note that this will be especially handy with presence of global register
allocator for (simple enough) functions where arguments can be kept in
registers.