RISC-V: do not allocate slot for return address if not needed
Up to now, RISC-V PSABI calling convention did not store return address
in frame if not needed (i.e., the function contains no "calls") but
still allocated the slot in frame. In other words, even if function
had no arguments and did not call anything (such as "return 42"), it
still allocated (and destroyed!) frame - unnecessarily.

This commit changes `#generatePrologue` and `generateEpilogue:` to
*not* generate code to allocate / destroy frame if not needed. resulting
it even smaller code for simple functions.