Fix bug in TRIL rewriting
Due to a bug, TRIL-rewriting optimization passed did not correctly
update all users of rewritten node. For example: imagine a pass wants
to rewrite all `iload x` with `iregload a0` in:

    ...
    N01  iadd
    N02    iload x
    N03    iconst 1
    ...
    N05  ireturn
    N02    ==> iload x

Resulting TRIL should look like:

    ...
    N01  iadd
    N88    iregload a0
    N03    iconst 1
    ...
    N05  ireturn
    N88    ==> iregload a0

Except that the second `iload x` was not rewritten (due to the bug).
This was because the guard in `TRILRewritingPass >> #evaluate:` that
return the *original* node if we have already seen it. Now, if the
*original* was replaced with *new*, the guard still returns the
original!

The fix is simple - instead of maintaining a simple set of seen nodes,
we keep a mapping containing all *original* nodes we have seen so far
to nodes that should replace them (if at all). This fixes the problem.

While at it, we maintain the mapping only per-basic block as to make
the dictionary smaller. Since nodes cannot be shared across block
boundary it is safe to do so.