Fix `TRILNodePrinter` to work with free-standing nodes
This commit fixes the printer to work with free-standing nodes, that is
with nodes that are not yet (or not anymore) anchored from a block. This
can happen temporarily for example when rewriting node (new node is
not yet inserted in place of the old one).

This fix enables debugging of such code (without random chokes in
printing, inspecting and so on).