Allow `glregdeps` node to be the last child of `bbstart`, `bbend` and branches
Like in previous commit, for global allocator to work it must be able
to express (global) register dependencies at CFG edges, that is, when
entering a block (`bbentry`) and leaving the block (`bbend` or any
branching node).

However, opcode specification does not explicitly say whether or not
a `glregdeps` can or cannot be a child in any way and the validation
code would reject it. Therefore this commit updates `#validateChildren:`
to allow `glregdeps` for some nodes.